# openapi/main.yaml
openapi: 3.0.0
info:
  title: Pooper API
  version: 1.0.0
  description: |
    API for the Pooper chat application.
    Provides endpoints for chat, MCP, tools, analytics, realtime, config, prompt, and preview services.
servers:
  - url: http://localhost:4100
    description: Local development server
...
components:
  responses:
    ApiError:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    ApiError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
paths:
...
  /api/config:
    get:
      tags:
        - Config
      summary: Get configuration
      description: Retrieve application configuration
      responses:
        '200':
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    put:
      tags:
        - Config
      summary: Update configuration
      description: Update application configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true

  # MCP routes
...
  /api/mcp/config:
    get:
      tags:
        - MCP
      summary: Get MCP configuration
      description: Retrieve MCP configuration
      responses:
        '200':
          description: MCP configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
    put:
      tags:
        - MCP
      summary: Update MCP configuration
      description: Update MCP configuration
      parameters:
        - name: createBackup
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Whether to create a backup of the current configuration before updating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: MCP configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  backupCreated:
                    type: boolean
                  backupPath:
                    type: string
        '400':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
...